src/services/forecast/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
requests==2.28.1
redis==4.5.4
python-decouple==3.8